pyproject.toml
[project]
name = "langchain-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.16",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.24",
]
